loader: Re-organize module layout
Re-organize modules of `loader` to follow image format, enable reuse of
code across different architectures. Its interface is unchanged.
