+++ src/lib.rs
@@ -68,7 +68,7 @@
 //!     let mut elf_pvh_image = Vec::new();
 //!     let path = concat!(
 //!         env!("CARGO_MANIFEST_DIR"),
-//!         "/src/loader/x86_64/elf/test_elfnote.bin"
+//!         "/src/loader/elf/test_elfnote.bin"
 //!     );
 //!     let mut file = File::open(path).unwrap();
 //!     file.read_to_end(&mut elf_pvh_image).unwrap();
