+++ src/loader/bzimage/fuzz_invalid_bzimage.bin
